chỉnh sửa chức năng thêm chi tiết toa thuốc, nếu thêm trùng thuốc 2 lần trở lên thì cộng dồn số lượng

+++ QuanLyPhongMachV3/src/main/java/com/dhb/springapp/controllers/admin/PrescriptionController.java
@@ -125,7 +125,8 @@ public class PrescriptionController {
             try {
                 ToaThuoc toaThuoc = iToaThuocService.getById(ToaThuoc.class, id);
                 HoaDon hoaDon = iHoaDonService.getHoaDonTheoToaThuoc(id);
-                iHoaDonService.delete(hoaDon);
+                if (hoaDon != null)
+                    iHoaDonService.delete(hoaDon);
                 iToaThuocService.delete(toaThuoc);
                 return "redirect:/admin/prescription";
             }
